[feat] 코드리뷰 반영 (세부적인거 + MatchResultCardView 컴포넌트로 분리)
