+++ SMASHING/Presentation/Home/MatchResultCreate/MatchResultCreateView.swift
@@ -15,7 +15,7 @@ final class MatchResultCreateView: BaseUIView {
     private var dropDownHeightConstraint: Constraint?
     private var isDropDownExpanded: Bool = false
     
-    private let navigationBar = CustomNavigationBar(title: "매쳥 결과 작성")
+    private let navigationBar = CustomNavigationBar(title: "결과 작성")
     
     private let titleLabel = UILabel().then {
         $0.text = "경기 결과를 작성해주세요"
@@ -29,67 +29,7 @@ final class MatchResultCreateView: BaseUIView {
         $0.textColor = .Text.tertiary
     }
     
-    private let containerView = UIView().then {
-        $0.backgroundColor = .Background.surface
-        $0.layer.cornerRadius = 12
-    }
-    
-    private let leftProfileStackView = UIStackView().then {
-        $0.axis = .vertical
-        $0.spacing = 0
-    }
-    
-    private let myImage = UIImageView().then {
-        $0.image = UIImage(systemName: "circle.fill")
-        $0.contentMode = .scaleAspectFit
-        $0.tintColor = .white
-    }
-    
-    private let myNickName = UILabel().then {
-        $0.text = "밤이달이"
-        $0.setPretendard(.textSmM)
-        $0.textColor = .Text.blue //muted 추가
-        $0.textAlignment = .center
-    }
-    
-    private let myScore = UILabel().then {
-        $0.text = "0"
-        $0.setPretendard(.headerHeroB)
-        $0.textColor = .Text.secondary
-        $0.textAlignment = .center
-    }
-    
-    private let scoreSemicolon = UILabel().then {
-        $0.text = ":"
-        $0.setPretendard(.headerHeroB)
-        $0.textColor = .Text.secondary
-        $0.textAlignment = .center
-    }
-    
-    private let rivalScore = UILabel().then {
-        $0.text = "0"
-        $0.setPretendard(.headerHeroB)
-        $0.textColor = .Text.secondary
-        $0.textAlignment = .center
-    }
-    
-    private let rightProfileStackView = UIStackView().then {
-        $0.axis = .vertical
-        $0.spacing = 0
-    }
-    
-    private let rivalImage = UIImageView().then {
-        $0.image = UIImage(systemName: "circle.fill")
-        $0.contentMode = .scaleAspectFit
-        $0.tintColor = .white
-    }
-    
-    private let rivalNickName = UILabel().then {
-        $0.text = "와구와구"
-        $0.setPretendard(.textSmM)
-        $0.textColor = .Text.blue //muted 추가
-        $0.textAlignment = .center
-    }
+    private let matchResultCard = MatchResultCardView()
     
     private let winnerLabel = UILabel().then {
         $0.text = "승자"
@@ -148,22 +88,16 @@ final class MatchResultCreateView: BaseUIView {
     let nextButton = CTAButton(label: "다음")
         
     override func setUI() {
-        leftProfileStackView.addArrangedSubviews(myImage, myNickName)
         
-        rightProfileStackView.addArrangedSubviews(rivalImage, rivalNickName)
         
-        containerView.addSubviews(leftProfileStackView,
-                                  rightProfileStackView,
-                                  myScore,
-                                  scoreSemicolon,
-                                  rivalScore)
+      
         
         dropDownOptionsView.addSubviews(myOptionButton, rivalOptionButton)
         
         addSubviews(navigationBar,
                     titleLabel,
                     subTitleLabel,
-                    containerView,
+                    matchResultCard,
                     winnerLabel,
                     scoreLabel,
                     winnerRequiredStar,
@@ -195,43 +129,9 @@ final class MatchResultCreateView: BaseUIView {
             $0.leading.trailing.equalToSuperview().inset(16)
         }
         
-        containerView.snp.makeConstraints {
-            $0.height.equalTo(160)
+        matchResultCard.snp.makeConstraints {
             $0.top.equalTo(subTitleLabel.snp.bottom).offset(28)
             $0.leading.trailing.equalToSuperview().inset(16)
-        }
-        
-        myImage.snp.makeConstraints {
-            $0.size.equalTo(64)
-        }
-        
-        leftProfileStackView.snp.makeConstraints {
-            $0.centerY.equalToSuperview()
-            $0.leading.equalTo(containerView.snp.leading).inset(38)
-        }
-        
-        myScore.snp.makeConstraints {
-            $0.trailing.equalTo(scoreSemicolon.snp.leading).offset(-2)
-            $0.centerY.equalTo(scoreSemicolon)
-        }
-        
-        scoreSemicolon.snp.makeConstraints {
-            $0.centerX.equalToSuperview()
-            $0.centerY.equalTo(myImage)
-        }
-        
-        rivalScore.snp.makeConstraints {
-            $0.leading.equalTo(scoreSemicolon.snp.trailing).offset(2)
-            $0.centerY.equalTo(scoreSemicolon)
-        }
-        
-        rivalImage.snp.makeConstraints {
-            $0.size.equalTo(64)
-        }
-        
-        rightProfileStackView.snp.makeConstraints {
-            $0.centerY.equalToSuperview()
-            $0.trailing.equalTo(containerView.snp.trailing).inset(38)
         }
         
         winnerLabel.snp.makeConstraints {
@@ -245,7 +145,7 @@ final class MatchResultCreateView: BaseUIView {
         }
         
         winnerDropDown.snp.makeConstraints {
-            $0.top.equalTo(containerView.snp.bottom).offset(20)
+            $0.top.equalTo(matchResultCard.snp.bottom).offset(20)
             $0.width.equalTo(140)
             $0.trailing.equalToSuperview().inset(16)
         }
@@ -328,13 +228,7 @@ final class MatchResultCreateView: BaseUIView {
         let rivalScore = rivalScoreTextField.text ?? "0"
         
 //        viewModel.updateScores(myScore: myScore, rivalScore: rivalScore)
-        self.myScore.text = "\(myScore)"
-        self.rivalScore.text = "\(rivalScore)"
+        
+        matchResultCard.updateScore(myScore: myScore, rivalScore: rivalScore)
     }
-}
-
-import SwiftUI
-@available(iOS 18.0, *)
-#Preview {
-    MatchResultCreateView()
 }
